+++ gems/shaolin-http/lib/shaolin/http.rb
@@ -3,6 +3,7 @@ require_relative "http/errors"
 require_relative "http/request"
 require_relative "http/rewindable_input"
 require_relative "http/controller"
+require_relative "http/open_api"
 require_relative "http/router"
 require_relative "http/provider"